Bump the wasm-tools crates (#3139)
* Bump the wasm-tools crates

Pulls in some updates here and there, mostly for updating crates to the
latest version to prepare for later memory64 work.

* Update lightbeam

+++ crates/lightbeam/src/microwasm.rs
@@ -440,7 +440,7 @@ impl From<WasmMemoryImmediate> for MemoryImmediate {
         assert_eq!(other.memory, 0);
         MemoryImmediate {
             flags: other.align.into(),
-            offset: other.offset,
+            offset: other.offset as u32,
         }
     }
 }
